+++ backend-rust/CHANGELOG.md
@@ -4,10 +4,24 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-Database schema version: 16
+## [0.1.38] - 2025-03-21
+
+Database schema version: 18
 
 ### Added
 
+- Add query `SuspendedValidators::suspendedValidators`.
+- Add query `SuspendedValidators::primedForSuspensionValidators`.
+- Add database migration 18 adding table tracking baker pool and passive stake for every reward period.
+- Add query `PassiveDelegation::apy` and `BakerPool::apy`.
+
+## [0.1.37] - 2025-03-21
+
+Database schema version: 17
+
+### Added
+
+- Add database migration 17 adding a table tracking reward metrics.
 - Add database migration 16 adding a table tracking commission rates for passive delegation and adding an index to retrieve passive delegators efficiently from the accounts table.
 - Add query `PassiveDelegation::delegators`, `PassiveDelegation::delegatorCount`,
 `PassiveDelegation::commissionRates`, `PassiveDelegation::delegatedStake` and `PassiveDelegation::delegatedStakePercentage`.